audit: T1/T2 conformance round two — 6 more defects, amendments 101-103
Three adversarial passes over the parser, the editor, and the three
implementations differentially. Every finding reproduced before acting.

Fixed:
- hex escapes were not ASCII: Character.digit is Unicode-aware and
  accepts 372 further code points, so a peer spelling \u in fullwidth
  or Arabic-Indic digits got content it never encoded — including a
  bare quote and a surrogate pair that then passed the pairing logic.
  The org.json class of laxity inside the parser written to close it
- Number.toInt() truncated wire positions to 32 bits: start 4294967296
  became 0, so a splice Emacs asked for OUTSIDE the text was applied at
  the head of the document and answered `applied` (16.1 forbids exactly
  this coercion). Now read as Long; out-of-domain fails the 19.1 range
  gate as a typed stale
- start + del overflowed Int, and a wrapped sum is not > n, so the
  range guard passed and substring got a negative index: one legal 4.2
  integer closed the transport
- a frame pipelined behind a malformed one was stranded until more
  bytes arrived — deadlock if the peer waits on that response
- edit.complete results were unvalidated: an ABSENT prefix became "",
  which selectCompletion's substring check trivially satisfied, so a
  malformed result inserted at the cursor without replacing
- the seed was never checked against max_editor_bytes (amendment 103)
- ebp.el matched Content-Length with ^/$, which in Emacs match at LINE
  boundaries; validate.py accepted lone surrogate escapes and had no
  unterminated-header cap

Report: docs/AUDIT-T1-T2-2026-07-25.md — also records 11 pre-existing
defects confirmed-but-not-fixed (phantom sessions from action args, no
session after reconnection, two surfaces sharing one editor identity,
LD-6 pendingEditors, key-based identity, offline save reaching the
durable queue, synchronized-editor drafts, dialogs bypassing the editor
limit, 19.5 annotations) and what was verified correct so it is not
re-litigated: 86,464 stream x chunking combinations through the
compiled decoder with zero divergence from the pre-rewrite decoder and
ebp.el, and 40,000 documents compared bit-for-bit against a C JSON
scanner.

wire 281 / app 20 / elisp 203 green.

+++ companion/wire/src/main/kotlin/com/calebc42/ebp/wire/EditorSession.kt
@@ -76,7 +76,10 @@ class EditorSession(
      */
     fun spliceJcsBytes(start: ScalarPos, del: Int, text: String): Long {
         val n = scalarLength()
-        if (start.v < 0 || del < 0 || start.v + del > n) return -1
+        // Long arithmetic: `start + del` as Int wraps, and a wrapped sum is
+        // NOT > n, so a legal SPEC 4.2 integer near Int.MAX_VALUE slipped
+        // this guard and reached substring() as a negative index.
+        if (start.v < 0 || del < 0 || start.v.toLong() + del > n) return -1
         val removed = shadow.substring(offset(start.v), offset(start.v + del))
         return jcsUtf8Bytes(shadow) - (jcsUtf8Bytes(removed) - 2) +
             (jcsUtf8Bytes(text) - 2)
@@ -91,7 +94,7 @@ class EditorSession(
      */
     fun splice(start: ScalarPos, del: Int, text: String, len: Int): Boolean {
         val n = scalarLength()
-        if (start.v < 0 || del < 0 || start.v + del > n) return false
+        if (start.v < 0 || del < 0 || start.v.toLong() + del > n) return false
         val inserted = text.codePointCount(0, text.length)
         if (len != n - del + inserted) return false
         shadow = shadow.substring(0, offset(start.v)) + text +
@@ -119,7 +122,7 @@ class EditorSession(
     fun spliceRemote(start: ScalarPos, del: Int, text: String, len: Int,
                      cursor: ScalarPos, selStart: ScalarPos?, selEnd: ScalarPos?): Boolean {
         val n = scalarLength()
-        if (start.v < 0 || del < 0 || start.v + del > n) return false
+        if (start.v < 0 || del < 0 || start.v.toLong() + del > n) return false
         val inserted = text.codePointCount(0, text.length)
         if (len != n - del + inserted) return false
         if (!caretValid(len, cursor.v, selStart?.v, selEnd?.v)) return false
